fix: ensure integer values for nperseg in PSD calculations across multiple scripts

+++ examples/GW150914_SMC.py
@@ -51,7 +51,7 @@ for ifo in ifos:
     ifo.set_data(data)
 
     psd_data = Data.from_gwosc(ifo.name, psd_start, psd_end)
-    ifo.set_psd(psd_data.to_psd(nperseg=data.duration * data.sampling_frequency))
+    ifo.set_psd(psd_data.to_psd(nperseg=int(data.duration * data.sampling_frequency)))
 
 # --- Waveform model ---
